requirements.txt
selenium
DrissionPage==3.2.34
matplotlib
opencv-python
apscheduler
bardapi==0.1.40
cachetools
azure-cognitiveservices-speech==1.28.0
Flask==2.2.2
PyYAML==6.0
requests==2.31.0
wechatpy==1.8.18
sseclient-py==1.7.2
Werkzeug==2.3.2
click==8.0
numpy
cryptography

paddlepaddle
paddleocr
fastapi
pyppeteer
google-generativeai
